SAML app with express Route extensions
