+++ pnpm-workspace.yaml
@@ -1,3 +1,4 @@
 packages:
   - "apps/mapp"
+  - "apps/saml"
   - "apps/xyz"
